ARP-2997: use pytest for running tests
